requirement.txt
torch>=1.7.0
torchvision>=0.8.1
opencv-python>=4.5.1
numpy>=1.19.5
Pillow>=8.1.0
scikit-learn>=0.24.1